+++ tsup.config.ts
@@ -6,7 +6,7 @@ export default defineConfig([
     format: ["esm"],
     sourcemap: true,
     clean: true,
-    target: "node20",
+    target: "node22",
     outDir: "dist",
     banner: {
       js: "#!/usr/bin/env node"
@@ -17,7 +17,7 @@ export default defineConfig([
     format: ["esm"],
     sourcemap: true,
     clean: false,
-    target: "node20",
+    target: "node22",
     outDir: "dist",
     dts: {
       compilerOptions: {